+++ src/plugins/rowSelector/components.ts
@@ -1,1 +1,2 @@
 export * from './basic/basicRowSelector.component';
+export * from './limited/limitedRowSelector.component';