+++ src/plugins/rowSelector/index.ts
@@ -1,2 +1,3 @@
 export * from './rowSelector.interface';
 export * from './basic/basicRowSelector.interface';
+export * from './limited/limitedRowSelector.interface';